requirements.txt
numpy==1.18.1
scikit-learn==1.0.2
scipy==1.5.2
pandas==1.2.4
Pillow==8.0.0
torchvision==0.10.0
torch==1.9.0
fastapi==0.75.0
python-multipart==0.0.5